Added image not found handler, moved checkbox to utils folder, cleaned up some conflicts, disabled several unnecessary rules in eslint
