+++ .eslintrc.js
@@ -20,5 +20,12 @@ module.exports = {
   rules: {
     'no-console': 'off',
     'react/prop-types': 'off',
+    'import/no-cycle': 'off',
+    'jsx-a11y/click-events-have-key-events': 'off',
+    'jsx-a11y/no-noninteractive-element-interactions': 'off',
+    'jsx-a11y/no-static-element-interactions': 'off',
+    'consistent-return': 'off',
+    'jsx-a11y/label-has-associated-control': 'off',
+    'no-alert': 'off',
   },
 };
