+++ client/src/assets/index.js
@@ -1,0 +1,3 @@
+const placeholderImage = 'https://i.ibb.co/HB10cH4/not-found.jpg';
+
+export default placeholderImage;
