+++ client/src/components/overview/components/Details/AddToCart.jsx
@@ -1,4 +1,3 @@
-/* eslint-disable no-alert */
 import React, { useEffect, useState } from 'react';
 
 function AddToCart({
