+++ client/src/components/overview/components/Details/StylePhoto.jsx
@@ -1,15 +1,12 @@
-/* eslint-disable jsx-a11y/no-noninteractive-element-interactions */
-/* eslint-disable jsx-a11y/click-events-have-key-events */
-/* eslint-disable jsx-a11y/no-static-element-interactions */
 import React, { useState } from 'react';
-import Checkbox from './Checkbox';
+import placeholderImage from '../../index';
+import Checkbox from '../utils/Checkbox';
 
 function StylePhoto({
   style, selectedStyle, setSelectedStyle, setSelectedPhoto, setSelectedThumb,
 }) {
   const photoSrc = style.photos[0].url;
   const thumbSrc = style.photos[0].thumbnail_url;
-  const placeholderImage = 'https://i.ibb.co/HB10cH4/not-found.jpg';
 
   const onImageError = (e) => {
     e.target.src = placeholderImage;
