+++ client/src/components/overview/components/Details/StyleSelector.jsx
@@ -1,6 +1,3 @@
-/* eslint-disable consistent-return */
-/* eslint-disable jsx-a11y/no-static-element-interactions */
-/* eslint-disable jsx-a11y/click-events-have-key-events */
 import React, { useEffect, useState } from 'react';
 import StylePhoto from './StylePhoto';
