+++ client/src/components/overview/index.jsx
@@ -1,11 +1,12 @@
+export { default as placeholderImage } from '../../assets/index';
 export { default as Basics } from './components/Basics/Basics';
 export { default as Price } from './components/Basics/Price';
 export { default as Description } from './components/Description';
 export { default as AddToCart } from './components/Details/AddToCart';
-export { default as Checkbox } from './components/Details/Checkbox';
 export { default as ProductDetails } from './components/Details/ProductDetails';
 export { default as QuantitySelector } from './components/Details/QuantitySelector';
 export { default as SizeSelector } from './components/Details/SizeSelector';
 export { default as StyleSelector } from './components/Details/StyleSelector';
-export { default as GalleryView } from './components/GalleryView';
+export { default as GalleryView } from './components/Gallery/GalleryView';
+export { default as Checkbox } from './components/utils/Checkbox';
 // root export file
